+++ src/main/java/lotto/strategy/LottoGenerateStrategy.java
@@ -1,4 +1,4 @@
-package lotto.core;
+package lotto.strategy;
 
 import java.util.List;
