+++ src/main/java/lotto/strategy/LottoWinningGenerateStrategy.java
@@ -1,4 +1,4 @@
-package lotto.core;
+package lotto.strategy;
 
 import java.util.Arrays;
 import java.util.List;
